pattern refactor
