+++ src/Pattern/Pattern.ts
@@ -1,38 +1,12 @@
-import { config } from './config';
-import { TCrossOrigin, TMat2D, TSize } from './typedefs';
-import { ifNaN } from './util/internals';
-import { uid } from './util/internals/uid';
-import { loadImage } from './util/misc/objectEnlive';
-import { pick } from './util/misc/pick';
-import { toFixed } from './util/misc/toFixed';
-import { classRegistry } from './ClassRegistry';
-
-export type TPatternRepeat = 'repeat' | 'repeat-x' | 'repeat-y' | 'no-repeat';
-
-type TExportedKeys =
-  | 'crossOrigin'
-  | 'offsetX'
-  | 'offsetY'
-  | 'patternTransform'
-  | 'repeat'
-  | 'source';
-
-export type PatternProps = Partial<Pick<Pattern, TExportedKeys>>;
-
-export type SerializedPatternProps = PatternProps & {
-  source: string;
-};
-
-export type TPatternHydrationOptions = {
-  /**
-   * handle aborting
-   * @see https://developer.mozilla.org/en-US/docs/Web/API/AbortController/signal
-   */
-  signal?: AbortSignal;
-};
-
-type TImageSource = { source: HTMLImageElement };
-type TCanvasSource = { source: HTMLCanvasElement };
+import { config } from '../config';
+import { TCrossOrigin, TMat2D } from '../typedefs';
+import { ifNaN } from '../util/internals';
+import { uid } from '../util/internals/uid';
+import { loadImage } from '../util/misc/objectEnlive';
+import { pick } from '../util/misc/pick';
+import { toFixed } from '../util/misc/toFixed';
+import { classRegistry } from '../ClassRegistry';
+import { TPatternRepeat, PatternProps } from './types';
 
 /**
  * @see {@link http://fabricjs.com/patterns demo}
@@ -119,7 +93,7 @@ export class Pattern {
   /**
    * @returns true if {@link source} is an <img> element
    */
-  isImageSource(): this is TImageSource {
+  isImageSource(): this is { source: HTMLImageElement } {
     return (
       !!this.source && typeof (this.source as HTMLImageElement).src === 'string'
     );
@@ -128,7 +102,7 @@ export class Pattern {
   /**
    * @returns true if {@link source} is a <canvas> element
    */
-  isCanvasSource(): this is TCanvasSource {
+  isCanvasSource(): this is { source: HTMLCanvasElement } {
     return !!this.source && !!(this.source as HTMLCanvasElement).toDataURL;
   }
