+++ src/Pattern/index.ts
@@ -1,0 +1,2 @@
+export { Pattern } from './Pattern';
+export * as Types from './types';
