+++ src/typedefs.ts
@@ -1,7 +1,7 @@
 // https://www.typescriptlang.org/docs/handbook/utility-types.html
 import { BaseFabricObject } from './EventTypeDefs';
 import type { Gradient } from './gradient/Gradient';
-import type { Pattern } from './Pattern';
+import type { Pattern } from './Pattern/Pattern';
 import type { XY, Point } from './Point';
 
 interface NominalTag<T> {
